+++ src/TooMany.WebClient/src/components/monitor/header/index.ts
@@ -1,0 +1,2 @@
+export { default as Tag } from './Tag';
+export { default as Task } from './Task';
