+++ src/TooMany.WebClient/src/components/monitor/header/title/Title.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from 'react';
+
+export default function ({ children }: { children: ReactNode }) {
+	return <h3 className="font-bold">{children}</h3>;
+}
