+++ src/TooMany.WebClient/src/components/monitor/header/title/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Title';
