+++ src/TooMany.WebClient/src/components/monitor/index.ts
@@ -1,0 +1,1 @@
+export * as Header from './header';
