+++ src/TooMany.WebClient/src/helpers/routes.ts
@@ -1,0 +1,11 @@
+export const home = () => '/';
+export const define = () => '/define';
+export const redefine = ({ name = ':name' }: { name?: string } = {}) =>
+	`/define/${name}`;
+export const monitor = ({
+	type = ':type',
+	name = ':name',
+}: {
+	type?: string;
+	name?: string;
+} = {}) => `/monitor/${type}/${name}`;
