+++ src/TooMany.WebClient/src/hooks/Navigation/index.ts
@@ -1,3 +1,2 @@
 export { default as useIsDefine } from './useIsDefine';
 export { default as useIsMonitor } from './useIsMonitor';
-export { default as useRoutes } from './useRoutes';
