+++ src/TooMany.WebClient/src/hooks/Navigation/useIsDefine.tsx
@@ -1,8 +1,7 @@
 import { useMatch } from 'react-router-dom';
-import useRoutes from './useRoutes';
+import * as routes from '@tm/helpers/routes';
 
 export default function () {
-	const routes = useRoutes();
 	const isDefine = useMatch(routes.define());
 	const isRedefine = useMatch(routes.redefine());
 	return isDefine || isRedefine;
