--- src/TooMany.WebClient/src/hooks/Navigation/useRoutes.tsx
@@ -1,14 +1,0 @@
-export default function () {
-	return {
-		home: () => '/',
-		define: () => '/define',
-		redefine: ({ name = ':name' }: { name?: string } = {}) => `/define/${name}`,
-		monitor: ({
-			type = ':type',
-			name = ':name',
-		}: {
-			type?: string;
-			name?: string;
-		} = {}) => `/monitor/${type}/${name}`,
-	};
-}
