+++ src/TooMany.WebClient/src/pages/Home.tsx
@@ -1,9 +1,8 @@
 import { memo } from 'react';
 import Link from '@components/link';
-import { useRoutes } from '@hooks/Navigation';
+import * as routes from '@tm/helpers/routes';
 
 function HomePage() {
-	const routes = useRoutes();
 	return (
 		<div className="flex min-h-screen min-w-full pl-52 pt-28">
 			<article>
